backend/requirements.txt
fastapi
uvicorn
pydantic
PyYAML
transformers
torch
accelerate
huggingface-hub
protobuf<4.0.0